+++ cytotable/apps/path.py
@@ -1,13 +1,14 @@
 """
-CytoTable: sources - tasks and flows related to
-source data and metadata for performing conversion work.
+cytotable.apps.path : work related to source data paths and metadata.
 """
-
+import logging
 import pathlib
 from typing import Any, Dict, List, Optional, Union
 
 from cloudpathlib import AnyPath
 from parsl.app.app import join_app, python_app
+
+logger = logging.getLogger(__name__)
 
 
 @python_app
@@ -44,7 +45,7 @@ def _build_path(
 
 
 @python_app
-def _get_source_filepaths(
+def _get_filepaths(
     path: Union[pathlib.Path, AnyPath],
     targets: List[str],
 ) -> Dict[str, List[Dict[str, Any]]]:
@@ -154,7 +155,7 @@ def _get_source_filepaths(
 
 
 @python_app
-def _infer_source_datatype(
+def _infer_path_datatype(
     sources: Dict[str, List[Dict[str, Any]]], source_datatype: Optional[str] = None
 ) -> str:
     """
@@ -235,47 +236,3 @@ def _filter_source_filepaths(
         ]
         for filegroup, files in sources.items()
     }
-
-
-@join_app
-def _gather_sources(
-    source_path: str,
-    source_datatype: Optional[str] = None,
-    targets: Optional[List[str]] = None,
-    **kwargs,
-) -> Dict[str, List[Dict[str, Any]]]:
-    """
-    Flow for gathering data sources for conversion
-
-    Args:
-        source_path: str:
-            Where to gather file-based data from.
-        source_datatype: Optional[str]:  (Default value = None)
-            The source datatype (extension) to use for reading the tables.
-        targets: Optional[List[str]]:  (Default value = None)
-            The source file names to target within the provided path.
-
-    Returns:
-        Dict[str, List[Dict[str, Any]]]
-            Data structure which groups related files based on the compartments.
-    """
-
-    from cytotable.sources import (
-        _build_path,
-        _filter_source_filepaths,
-        _get_source_filepaths,
-        _infer_source_datatype,
-    )
-
-    source_path = _build_path(path=source_path, **kwargs)
-
-    # gather filepaths which will be used as the basis for this work
-    sources = _get_source_filepaths(path=source_path, targets=targets)
-
-    # infer or validate the source datatype based on source filepaths
-    source_datatype = _infer_source_datatype(
-        sources=sources, source_datatype=source_datatype
-    )
-
-    # filter source filepaths to inferred or source datatype
-    return _filter_source_filepaths(sources=sources, source_datatype=source_datatype)
