+++ tests/test_convert.py
@@ -20,18 +20,18 @@ from parsl.providers import LocalProvider
 from pyarrow import csv, parquet
 from pycytominer.cyto_utils.cells import SingleCells
 
-from cytotable.convert import (
+from cytotable.apps.combine import (
     _concat_join_sources,
     _concat_source_group,
     _get_join_chunks,
     _infer_source_group_common_schema,
     _join_source_chunk,
-    _prepend_column_name,
-    _to_parquet,
-    convert,
 )
+from cytotable.apps.modify import _prepend_column_name
+from cytotable.apps.path import _get_filepaths, _infer_path_datatype
+from cytotable.apps.workflow import _to_parquet
+from cytotable.convert import convert
 from cytotable.presets import config
-from cytotable.sources import _get_source_filepaths, _infer_source_datatype
 from cytotable.utils import (
     _column_sort,
     _duckdb_reader,
@@ -59,22 +59,22 @@ def test_config():
         ) == sorted(config_preset.keys())
 
 
-def test_get_source_filepaths(get_tempdir: str, data_dir_cellprofiler: str):
+def test_get_filepaths(get_tempdir: str, data_dir_cellprofiler: str):
     """
-    Tests _get_source_filepaths
+    Tests _get_filepaths
     """
 
     # test that no sources raises an exception
     empty_dir = pathlib.Path(f"{get_tempdir}/temp")
     empty_dir.mkdir(parents=True, exist_ok=True)
     with pytest.raises(Exception):
-        single_dir_result = _get_source_filepaths(
+        single_dir_result = _get_filepaths(
             path=empty_dir,
             targets=["image", "cells", "nuclei", "cytoplasm"],
         ).result()
 
     # check that single sqlite file is returned as desired
-    single_file_result = _get_source_filepaths(
+    single_file_result = _get_filepaths(
         path=pathlib.Path(
             f"{data_dir_cellprofiler}/NF1_SchwannCell_data/all_cellprofiler.sqlite"
         ),
@@ -83,20 +83,20 @@ def test_get_source_filepaths(get_tempdir: str, data_dir_cellprofiler: str):
     assert len(set(single_file_result.keys())) == 1
 
     # check that single csv file is returned as desired
-    single_file_result = _get_source_filepaths(
+    single_file_result = _get_filepaths(
         path=pathlib.Path(f"{data_dir_cellprofiler}/ExampleHuman/Cells.csv"),
         targets=["cells"],
     ).result()
     assert len(set(single_file_result.keys())) == 1
 
-    single_dir_result = _get_source_filepaths(
+    single_dir_result = _get_filepaths(
         path=pathlib.Path(f"{data_dir_cellprofiler}/ExampleHuman"),
         targets=["cells"],
     ).result()
     # test that the single dir structure includes 1 unique key (for cells)
     assert len(set(single_dir_result.keys())) == 1
 
-    single_dir_result = _get_source_filepaths(
+    single_dir_result = _get_filepaths(
         path=pathlib.Path(f"{data_dir_cellprofiler}/ExampleHuman"),
         targets=["image", "cells", "nuclei", "cytoplasm"],
     ).result()
@@ -424,26 +424,26 @@ def test_concat_join_sources(get_tempdir: str):
     )
 
 
-def test_infer_source_datatype():
+def test_infer_path_datatype():
     """
-    Tests _infer_source_datatype
+    Tests _infer_path_datatype
     """
 
     data = {
         "sample_1.csv": [{"source_path": "stub"}],
         "sample_2.CSV": [{"source_path": "stub"}],
     }
-    assert _infer_source_datatype(sources=data).result() == "csv"
+    assert _infer_path_datatype(sources=data).result() == "csv"
     with pytest.raises(Exception):
-        _infer_source_datatype(sources=data, source_datatype="parquet").result()
+        _infer_path_datatype(sources=data, source_datatype="parquet").result()
 
     data["sample_3.parquet"] = [{"source_path": "stub"}]
     assert (
-        _infer_source_datatype(sources=data, source_datatype="parquet").result()
+        _infer_path_datatype(sources=data, source_datatype="parquet").result()
         == "parquet"
     )
     with pytest.raises(Exception):
-        _infer_source_datatype(sources=data).result()
+        _infer_path_datatype(sources=data).result()
 
 
 def test_to_parquet(
